+++ db/migrate/20130822191157_add_randomization_goal_to_projects.rb
@@ -1,0 +1,5 @@
+class AddRandomizationGoalToProjects < ActiveRecord::Migration
+  def change
+    add_column :projects, :randomization_goal, :integer, null: false, default: 0
+  end
+end