+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Lists are now generated as a product of the stratification factor options
   - Randomization lists are now popuplated by shuffled blocks
   - Different block sizes and allocations can now be specified
+  - Randomization goal can now be specified
 
 ## 0.0.0 (August 21, 2013)